requirements.txt
# Core dependencies for RAG pipeline
torch>=2.0.0
sentence-transformers>=2.2.2
transformers>=4.36.0
langchain>=0.1.0
chromadb>=0.4.0
pandas
numpy
PyMuPDF

# OpenAI API for LLM
openai>=1.0.0

# Environment variables management
python-dotenv>=1.0.0

# Optional: For better progress tracking
tqdm

# Optional: For API development (if building web interface)
# fastapi>=0.104.0
# uvicorn>=0.24.0
